pyproject.toml
[project]
name = "dspy-sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy>=3.1.3",
    "python-dotenv>=1.2.1",
]
